Relocating shared types for fx usage
